Fix edit file description bug

+++ htdocs/artefact/file/updatemetadata.json.php
@@ -36,15 +36,16 @@ $name = param_variable('name');
 $description = param_variable('description');
 $collideaction = param_variable('collideaction', 'fail');
 
-
 if ($existingid = ArtefactTypeFileBase::exists_in_db($name, $USER->get('id'), $parentfolder)) {
-    log_debug($existingid);
-    if ($collideaction == 'replace') {
-        $copy = artefact_instance_from_id($existingid);
-        $copy->delete();
-    }
-    else {
-        json_reply('local', get_string('fileexists', 'artefact.file'));
+    if ($existingid != $id) {
+        if ($collideaction == 'replace') {
+            log_debug('deleting ' . $existingid);
+            $copy = artefact_instance_from_id($existingid);
+            $copy->delete();
+        }
+        else {
+            json_reply('local', get_string('fileexists', 'artefact.file'));
+        }
     }
 }
